pyproject.toml
[project]
name = "slack_exporter"
version = "0.1.0"
description = ""
authors = [
    {name = "Hubert Fauconnier",email = "h.fauconnier@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-api-python-client (>=2.174.0,<3.0.0)",
    "isort (>=6.0.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
export_slack_to_mega = "slack_exporter.main:export_slack_to_mega"
export_slack_to_googledrive = "slack_exporter.main:export_slack_to_googledrive"
